# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: YOGA LIFE VN API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (super admin only)
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [student, center_admin, super_admin]
        - name: centerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /memberships:
    get:
      operationId: listMemberships
      tags: [memberships]
      summary: List membership plans
      parameters:
        - name: centerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of membership plans
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembershipsListResponse"
    post:
      operationId: createMembership
      tags: [memberships]
      summary: Create membership plan (center admin/super admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMembershipRequest"
      responses:
        "201":
          description: Created membership plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"
  /memberships/{id}:
    get:
      operationId: getMembership
      tags: [memberships]
      summary: Get membership plan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Membership details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"
    put:
      operationId: updateMembership
      tags: [memberships]
      summary: Update membership plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMembershipRequest"
      responses:
        "200":
          description: Updated membership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"
    delete:
      operationId: deleteMembership
      tags: [memberships]
      summary: Delete membership plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [student, center_admin, super_admin]
        centerId:
          type: integer
          nullable: true
        centerName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, email, fullName, role, createdAt]
    UpdateUserRequest:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [student, center_admin, super_admin]
        centerId:
          type: integer
          nullable: true
        avatarUrl:
          type: string
          nullable: true
    UsersListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
      required: [users, total]
...
    Membership:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        durationDays:
          type: integer
        classesPerWeek:
          type: integer
          nullable: true
        centerId:
          type: integer
        centerName:
          type: string
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, price, durationDays, centerId, centerName, features, isActive, createdAt]
    CreateMembershipRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        durationDays:
          type: integer
        classesPerWeek:
          type: integer
          nullable: true
        centerId:
          type: integer
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
      required: [name, price, durationDays, centerId, features]
    UpdateMembershipRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        durationDays:
          type: integer
        classesPerWeek:
          type: integer
          nullable: true
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
    MembershipsListResponse:
      type: object
      properties:
        memberships:
          type: array
          items:
            $ref: "#/components/schemas/Membership"
        total:
          type: integer
      required: [memberships, total]
...